refactor: localize worker data contracts
